% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{getAudioDevices}
\alias{getAudioDevices}
\title{Get List of Audio Devices}
\usage{
getAudioDevices()
}
\value{
a character vector
}
\description{
return a character vector containg names of audio devices.
note that you may successfully open the default device even if you fail to get a list of audio devices.
}
