% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{human}
\alias{human}
\title{Drawing of Human}
\format{
an array
}
\description{
The drawing of human was drawn by Shota Ochi.
It is licensed under \href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 Universal}.
}
