% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{pick}
\title{Pick points, lines, or rectangles from image}
\usage{
pick(
  img,
  type = "point",
  n = 1,
  mark = TRUE,
  color = c(0, 0, 0, 255),
  pos = "undefined",
  chorder = "cxy",
  title = "",
  renderer_idx = -1
)
}
\arguments{
\item{img}{an array that represents image.}

\item{type}{"point", "line", or "rectangle".}

\item{n}{the number of points, lines, or rectangles to be picked.}

\item{mark}{whether to mark or not.}

\item{color}{a vector that represents a color (c(R, G, B, A)).}

\item{pos}{position of window. a numeric vector that represents position, "centered", or "undefined".}

\item{chorder}{channel order of image. "cxy" or "xyc".}

\item{title}{title of window.}

\item{renderer_idx}{the number of rendering device to be initialized.}
}
\value{
a data.frame. note that x and y are the coordinates of upper left corner of rectangle when type is "rect".
}
\description{
Pick points, lines, or rectangles from image. note that the origin is top left.
}
\examples{
if (interactive())
{
  pick(human)
  pick(human, type = "line")
  pick(human, type = "rect")
}
}
