% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{SDLaudio}
\alias{SDLaudio}
\title{R6 class SDL Audio Interaface}
\description{
R6 class SDL Audio Interaface

R6 class SDL Audio Interaface
}
\details{
SDLaudio
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{extptr}}{external pointer of SDL audio.}

\item{\code{path}}{to a wav file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SDLaudio-new}{\code{SDLaudio$new()}}
\item \href{#method-SDLaudio-clear}{\code{SDLaudio$clear()}}
\item \href{#method-SDLaudio-getstatus}{\code{SDLaudio$getstatus()}}
\item \href{#method-SDLaudio-pause}{\code{SDLaudio$pause()}}
\item \href{#method-SDLaudio-unpause}{\code{SDLaudio$unpause()}}
\item \href{#method-SDLaudio-queue}{\code{SDLaudio$queue()}}
\item \href{#method-SDLaudio-clone}{\code{SDLaudio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDLaudio-new"></a>}}
\if{latex}{\out{\hypertarget{method-SDLaudio-new}{}}}
\subsection{Method \code{new()}}{
create a SDLaudio object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDLaudio$new(path, dev)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path to a wav file.}

\item{\code{dev}}{audio device name to use. can be missing.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new SDLaudio object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDLaudio-clear"></a>}}
\if{latex}{\out{\hypertarget{method-SDLaudio-clear}{}}}
\subsection{Method \code{clear()}}{
clear queued audio.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDLaudio$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDLaudio-getstatus"></a>}}
\if{latex}{\out{\hypertarget{method-SDLaudio-getstatus}{}}}
\subsection{Method \code{getstatus()}}{
get status of audio device.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDLaudio$getstatus()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDLaudio-pause"></a>}}
\if{latex}{\out{\hypertarget{method-SDLaudio-pause}{}}}
\subsection{Method \code{pause()}}{
pause audio playback.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDLaudio$pause()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDLaudio-unpause"></a>}}
\if{latex}{\out{\hypertarget{method-SDLaudio-unpause}{}}}
\subsection{Method \code{unpause()}}{
unpause audio playback.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDLaudio$unpause()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDLaudio-queue"></a>}}
\if{latex}{\out{\hypertarget{method-SDLaudio-queue}{}}}
\subsection{Method \code{queue()}}{
queue audio.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDLaudio$queue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SDLaudio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SDLaudio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDLaudio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
