% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{train-whistle.wav}
\alias{train-whistle.wav}
\title{Sound File of Train Whistle}
\description{
The original file was downloaded from \url{https://freesound.org/people/RetroGuy23/sounds/652401/download/652401__retroguy23__train-whistle.mp3}.
The sound "Train Whistle" by RetroGuy23 is licensed under CC BY-NC 4.0 (\url{https://creativecommons.org/licenses/by-nc/4.0/}).
The original file was converted into wav file.
}
\examples{
system.file("extdata", "train-whistle.wav", package = "sdlr")
}
