% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{Display an image.}
\usage{
view(img, pos = "undefined", chorder = "cxy", title = "", renderer_idx = -1)
}
\arguments{
\item{img}{an array that represents an image.}

\item{pos}{position of window. a numeric vector that represents position, "centered", or "undefined".}

\item{chorder}{channel order of image. "cxy" or "xyc".}

\item{title}{title of window.}

\item{renderer_idx}{number of rendering device to be initialized.}
}
\description{
Display an image that is represented by a numeric array or a raw array.
}
\examples{
if (interactive())
{
  view(human)
}
}
